# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CIMA CRM - mod-media
  version: 1.0.0
  description: |
    Microservicio de media para avatares y documentos. Expone contratos directos
    del servicio; el gateway puede publicarlos como proxy sin fusionar specs.
servers:
  - url: http://localhost:3002
    description: mod-media directo (desarrollo)
tags:
  - name: Sistema
  - name: Documentacion
  - name: Avatares
  - name: Documentos
components:
  securitySchemes:
    GatewayIdentity:
      type: apiKey
      in: header
      name: X-User-Id
      description: Headers de identidad propagados por el gateway confiable.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Avatar:
      type: object
      properties:
        userId:
          type: string
        url:
          type: string
          format: uri
        objectKey:
          type: string
        updatedAt:
          type: string
          format: date-time
    DocumentUploadUrl:
      type: object
      required: [uploadUrl, objectKey, expiresInSeconds]
      properties:
        uploadUrl:
          type: string
          format: uri
        objectKey:
          type: string
        expiresInSeconds:
          type: integer
    DocumentAccess:
      type: object
      required: [url, expiresInSeconds]
      properties:
        url:
          type: string
          format: uri
        expiresInSeconds:
          type: integer
    DataEnvelope:
      type: object
      properties:
        data:
          type: object
security:
  - GatewayIdentity: []
paths:
  /health:
    get:
      tags: [Sistema]
      summary: Health check
      security: []
      responses:
        "200":
          description: Servicio activo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: mod-media
  /openapi.yaml:
    get:
      tags: [Documentacion]
      summary: Especificacion OpenAPI del servicio
      security: []
      responses:
        "200":
          description: Archivo OpenAPI YAML
          content:
            application/yaml:
              schema:
                type: string
  /media/avatars:
    post:
      tags: [Avatares]
      summary: Subir avatar del usuario actual
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Avatar actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DataEnvelope"
  /media/avatars/current:
    get:
      tags: [Avatares]
      summary: Obtener avatar vigente del usuario actual
      responses:
        "200":
          description: Avatar vigente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DataEnvelope"
  /media/avatars/users:
    get:
      tags: [Avatares]
      summary: Obtener avatares vigentes de varios usuarios
      parameters:
        - in: query
          name: ids
          schema:
            type: string
          description: Lista de IDs separada por coma.
      responses:
        "200":
          description: Avatares por usuario
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DataEnvelope"
  /media/documents/upload-url:
    post:
      tags: [Documentos]
      summary: Generar URL prefirmada de carga para documentos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, mimeType, sizeBytes]
              properties:
                fileName:
                  type: string
                mimeType:
                  type: string
                sizeBytes:
                  type: integer
      responses:
        "200":
          description: URL de carga generada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DocumentUploadUrl"
  /media/documents/confirm:
    post:
      tags: [Documentos]
      summary: Confirmar carga directa de documento en OCI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectKey, fileName, mimeType, sizeBytes]
              properties:
                objectKey:
                  type: string
                fileName:
                  type: string
                mimeType:
                  type: string
                sizeBytes:
                  type: integer
      responses:
        "201":
          description: Documento registrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DataEnvelope"
  /media/documents/access:
    get:
      tags: [Documentos]
      summary: Generar URL prefirmada de lectura para un documento
      parameters:
        - in: query
          name: objectKey
          required: true
          schema:
            type: string
        - in: query
          name: download
          schema:
            type: boolean
      responses:
        "200":
          description: URL de acceso generada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DocumentAccess"
  /media/documents:
    delete:
      tags: [Documentos]
      summary: Eliminar documento por objectKey
      parameters:
        - in: query
          name: objectKey
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documento eliminado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DataEnvelope"